+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-spring.application.name=xml-retrieval-service
-server.port=8110
+server.servlet.context-path=/courtlist
+server.port=8083
