+++ src/test/java/ca/bc/gov/open/CourtlistApiApplicationTests.java
@@ -1,10 +1,10 @@
-package ca.bc.gov.restxml;
+package ca.bc.gov.open;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class RestXmlApplicationTests {
+class CourtlistApiApplicationTests {
 
 	@Test
 	void contextLoads() {
